NS-100/chart-managers: Rename chart -> response
