+++ net-gateway/src/endpoints/charts/network_bandwidth_per_endpoint/mod.rs
@@ -1,3 +1,3 @@
-pub mod chart;
+pub mod response;
 pub mod endpoint;
 pub mod request;
