+++ net-gateway/src/endpoints/charts/network_bandwidth_per_endpoint/response/mod.rs
